03_Token_service/requirement_tokent_service.txt
Django==2.2.4
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-oidc-provider==0.7.0
django-python3-ldap==0.11.2
django-registration==3.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
jsonschema==2.6.0
requests==2.18.4
reverse==0.1.0
oauthlib==2.1.0
django-oauth-toolkit==1.2.0
cryptography==2.3.1
pyjwkest==1.4.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
urllib3==1.22
pip==19.2.3